fixed urls, added comments, fixed login issue for favorites, merged normal and parallel version

+++ wallhaven.sh
@@ -8,6 +8,12 @@
 # This Script is written for GNU Linux, it should work under Mac OS
 #
 #
+# Revision 0.1.2
+# 1. fixed urls to work with latest wallhaven update
+# 2. added some comments
+# 3. fixed login issue when downloading favorites
+# 4. merged normal and parallel version
+# 
 # Revision 0.1.1
 # 1. updated and tested parts of the script to work with 
 #    newest wallhaven site (not all features tested)
@@ -16,46 +22,48 @@
 # 1. first Version of script, most features from the wallbase 
 #    script are implemented
 #
-##################################
-###    Needed for NSFW/New     ###
-##################################
+#####################################
+###   Needed for NSFW/Favorites   ###
+#####################################
 # Enter your Username
 USER=""
 # Enter your password
 PASS=""
-#################################
-###  End needed for NSFW      ###
-#################################
+#####################################
+### End needed for NSFW/Favorites ###
+#####################################
 
 
-#################################
-###   Configuration Options   ###
-#################################
+####################################
+###     Configuration Options    ###
+####################################
 # Where should the Wallpapers be stored?
 LOCATION=/location/to/your/wallpaper/folder
-# How many Wallpapers should be downloaded, should be multiples of 64 (right now they only use a fixed number of thumbs per page)
-WPNUMBER=64
-# Type standard (newest, oldest, random, hits, mostfav), search, favorites, useruploads
+# How many Wallpapers should be downloaded, should be multiples of 24 (right now they only use a fixed number of thumbs per page)
+WPNUMBER=48
+# Type standard (newest, oldest, random, hits, mostfav), search, favorites (for now only the default collection), useruploads (if selected, only purity variable will change the outcome)
 TYPE=standard
-# From which Categories should Wallpapers be downloaded
+# From which Categories should Wallpapers be downloaded, first number is for General, second for Anime, third for People, 1 to enable category, 0 to disable it
 CATEGORIES=111
-# Which Purity Wallpapers should be downloaded
+# Which Purity Wallpapers should be downloaded, first number is for sfw content, second for sketchy content, third for nsfw, 1 to enable, 0 to disable
 PURITY=110
-# Which Resolution should be downloaded, leave empty for all
+# Which Resolution should be downloaded, leave empty for all (most common resolutions possible, for details see wallhaven site)
 RESOLUTION=
-# Which aspectratio should be downloaded, leave empty for all
+# Which aspectratio should be downloaded, leave empty for all (possible values: 4x3, 5x4, 16x9, 16x10, 32x9, 48x9)
 RATIO=
 # Which Type should be displayed (relevance, random, date_added, views, favorites)
 SORTING=random
 # How should the Wallpapers be ordered (desc, asc)
 ORDER=desc
-# Searchterm
+# Searchterm, only used if TYPE = search
 QUERY="nature"
 # User from which Wallpapers should be downloaded (only used for TYPE=useruploads)
 USR=AksumkA
-#################################
-### End Configuration Options ###
-#################################
+# use gnu parallel to speed up the download (0, 1), if set to 1 make sure you have gnuparallel installed, see normal.vs.parallel.txt for speed improvements
+PARALLEL=0
+####################################
+###   End Configuration Options  ###
+####################################
  
 if [ ! -d $LOCATION ]; then
     mkdir -p $LOCATION
@@ -107,7 +115,7 @@ function getPage {
 # arg1: the file containing the wallpapers
 #
 function downloadWallpapers {
-	URLSFORIMAGES="$(cat tmp | grep -o '<a href="http://alpha.wallhaven.cc/wallpaper/[0-9]*"' | sed  's .\{25\}  ')"
+	URLSFORIMAGES="$(cat tmp | grep -o '<a class="preview" href="http://alpha.wallhaven.cc/wallpaper/[0-9]*"' | sed  's .\{25\}  ')"
 	for imgURL in $URLSFORIMAGES
         do
         img="$(echo $imgURL | sed 's/.\{1\}$//')"
@@ -115,26 +123,38 @@ function downloadWallpapers {
         if cat downloaded.txt | grep -w "$number" >/dev/null
 			then
 				printf "File already downloaded!\n"
-			else
-				echo $number >> downloaded.txt
-				wget -q --keep-session-cookies --load-cookies=cookies.txt --referer=alpha.wallhaven.cc $img
-				cat $number | egrep -o "http://alpha.wallhaven.cc/wallpapers.*(png|jpg|gif)" | wget -q --keep-session-cookies --load-cookies=cookies.txt --referer=http://alpha.wallhaven.cc/wallpaper/$number -i -
-				rm $number	
-        fi
+        elif [ $PARALLEL == 1 ]
+            then
+                echo $number >> downloaded.txt
+                echo $number >> download.txt
+        else
+                echo $number >> downloaded.txt
+				  wget -q --keep-session-cookies --load-cookies=cookies.txt --referer=alpha.wallhaven.cc $img
+                cat $number | egrep -o 'wallpapers.*(png|jpg|gif)' | wget -q --keep-session-cookies --load-cookies=cookies.txt --referer=http://alpha.wallhaven.cc/wallpaper/$number -i -
+                rm $number
+            fi
         done
-        rm tmp
+
+    if [ $PARALLEL == 1 ]
+        then
+            cat download.txt | parallel --no-notice wget -q --keep-session-cookies --load-cookies=cookies.txt --referer=alpha.wallhaven.cc http://alpha.wallhaven.cc/wallpaper/{}
+            cat download.txt | parallel --no-notice "cat {} | egrep -o 'wallpapers.*(png|jpg|gif)' | wget -q --keep-session-cookies --load-cookies=cookies.txt --referer=http://alpha.wallhaven.cc/wallpaper/{} -i -"
+            rm tmp $(cat download.txt) download.txt
+        else
+            rm tmp
+        fi
 } #/downloadWallpapers
 
 # login only when it is required ( for example to download favourites or nsfw content... )
-if [ $PURITY == 001 ] || [ $PURITY == 011 ] || [ $PURITY == 111 ] ; then
+if [ $PURITY == 001 ] || [ $PURITY == 011 ] || [ $PURITY == 111 ] || [ $TYPE == favorites ] ; then
    login $USER $PASS
 fi
 
 if [ $TYPE == standard ]; then
-    for (( count= 0, page=1; count< "$WPNUMBER"; count=count+64, page=page+1 ));
+    for (( count=0, page=1; count< "$WPNUMBER"; count=count+24, page=page+1 ));
     do
         printf "Download Page $page"
-        getPage "wallpaper/search?page=$page&categories=$CATEGORIES&purity=$PURITY&resolutions=$RESOLUTION&ratios=$RATIO&sorting=$SORTING&order=$ORDER"
+        getPage "search?page=$page&categories=$CATEGORIES&purity=$PURITY&resolutions=$RESOLUTION&ratios=$RATIO&sorting=$SORTING&order=$ORDER"
         printf "                    - done!\n"
         printf "Download Wallpapers from Page $page"
         downloadWallpapers
@@ -143,10 +163,10 @@ if [ $TYPE == standard ]; then
 
 elif [ $TYPE == search ] ; then
     # SEARCH
-    for (( count= 0, page=1; count< "$WPNUMBER"; count=count+"64", page=page+1 ));
+    for (( count=0, page=1; count< "$WPNUMBER"; count=count+24, page=page+1 ));
     do
         printf "Download Page $page"
-        getPage "wallpaper/search?page=$page&categories=$CATEGORIES&purity=$PURITY&resolutions=$RESOLUTION&ratios=$RATIO&sorting=relevance&order=desc&q=$QUERY"
+        getPage "search?page=$page&categories=$CATEGORIES&purity=$PURITY&resolutions=$RESOLUTION&ratios=$RATIO&sorting=relevance&order=desc&q=$QUERY"
         printf "                    - done!\n"
         printf "Download Wallpapers from Page $page"
         downloadWallpapers
@@ -156,8 +176,8 @@ elif [ $TYPE == search ] ; then
 elif [ $TYPE == favorites ] ; then
     # FAVORITES
     # currently using sum of all collections
-    favnumber="$(wget -q --keep-session-cookies --load-cookies=cookies.txt --referer=alpha.wallhaven.cc http://alpha.wallhaven.cc/favorites -O - | grep -A 1 "<span>Favorites</span>" | grep -B 1 "<small>" | sed -n '2{p;q}' | sed 's/.\{9\}$//' | sed 's .\{23\}  ')"
-    for (( count= 0, page=1; count< "$WPNUMBER" && count< "$favnumber"; count=count+"64", page=page+1 ));
+    favnumber="$(wget -q --keep-session-cookies --load-cookies=cookies.txt --referer=alpha.wallhaven.cc http://alpha.wallhaven.cc/favorites -O - | grep -A 1 "<span>Favorites</span>" | grep -B 1 "<small>" | sed -n '2{p;q}' | sed 's/<[^>]\+>/ /g')"
+    for (( count=0, page=1; count< "$WPNUMBER" && count< "$favnumber"; count=count+64, page=page+1 ));
     do
         printf "Download Page $page"
         getPage "favorites?page=$page"
@@ -169,10 +189,10 @@ elif [ $TYPE == favorites ] ; then
 
 elif [ $TYPE == useruploads ] ; then
     # UPLOADS FROM SPECIFIC USER
-    for (( count= 0, page=1; count< "$WPNUMBER"; count=count+"64", page=page+1 ));
+    for (( count=0, page=1; count< "$WPNUMBER"; count=count+24, page=page+1 ));
     do
         printf "Download Page $page"
-        getPage "user/$USR/uploads?page=$page"
+        getPage "user/$USR/uploads?page=$page&purity=$PURITY"
         printf "                    - done!\n"
         printf "Download Wallpapers from Page $page"
         downloadWallpapers
